+++ web/typescript/shared/MotionInitiatorShow.ts
@@ -1,6 +1,6 @@
 class MotionInitiatorShow {
     constructor() {
-        $(".motionData .contactShow").click(function (ev) {
+        $(".motionData .contactShow").on("click", function (ev) {
             ev.preventDefault();
             $(this).addClass("hidden");
             $(".motionData .contactDetails").removeClass("hidden");