+++ History.md
@@ -8,6 +8,7 @@
 - Amendments can now have initiator and supporter settings independent from the base motions.
 - A slightly new homepage layout variant is introduced, which first shows the agenda, then the motions, but by default hides the amendments below the motions, with the option to toggle the list. This is meant for consultation with hundreds of amendments where the sheer amount of amendments makes the list way too long.
 - When merging amendments into motions, conflicting text passages can now be marked as "handled" and therefore hidden.
+- Custom themes can now modify the font size of the headings.
 - Bugfix: When merging amendments with a modified version as proposed procedure, the paragraph's collision section was initially showing the collisions of the original amendment, not the proposed version.
 
 ### Version 4.3.1 (2019-10-27)